# src/app.py
# TODO
# Check bei Post Methode ob die Datenstruktur passt
# Fehlermeldungen (Response usw.)
# Doku anzeigen lassen (swagger.json oder yaml)
# eigenen Ref code erstellen falls noch nicht vorhanden
# anstatts lokalhost -> api.fiat2defi.ch/


import ssl
import pymongo

from flask_swagger import swagger
from bson.json_util import dumps
from flask import abort
from flask import request, Flask, jsonify
from flask_swagger_ui import get_swaggerui_blueprint

app = Flask(__name__)

# Connect to mongoDB
client = pymongo.MongoClient(
    "mongodb+srv://apiuser:apiuser@defiexchange.cihof.mongodb.net/defiexchange?retryWrites=true&w=majority",
    ssl_cert_reqs=ssl.CERT_NONE)

app.run()

# Get user information
@app.route('/api/v1/userInformation', methods=['GET'])
def getUserInformation():
    """Returns all user information from legacy address"""
    query_parameters = request.args
    legacyAddress = query_parameters.get('legacyAddress')
    db = client['defiexchange']
    coll = db['customers']
    return dumps(coll.find_one({"address": legacyAddress}), indent=2)


# Get account history
@app.route('/api/v1/accountHistory', methods=['GET'])
def getTransactionsHistory():
    query_parameters = request.args
    legacyAddress = query_parameters.get('legacyAddress')
    db = client['defiexchange']
    trans = db['transactions']
    return dumps(trans.find_one({"address": legacyAddress}), indent=2)


# Get all user information
@app.route('/api/v1/allUserInformation', methods=['GET'])
def getAllUserInformation():
    query_parameters = request.args
    auth = query_parameters.get('Auth')
    db = client['defiexchange']
    trans = db['admin']
    if trans.find_one({"oAuth": auth}) is not None:
        db = client['defiexchange']
        coll = db['customers']
        return dumps(coll.find({}), indent=2)
    else:
        abort(400)


# Get all account history
@app.route('/api/v1/allAccountHistory', methods=['GET'])
def getAllTransactionsHistory():
    query_parameters = request.args
    auth = query_parameters.get('Auth')
    db = client['defiexchange']
    trans = db['admin']
    if trans.find_one({"oAuth": auth}) is not None:
        db = client['defiexchange']
        coll = db['transactions']
        return dumps(coll.find({}), indent=2)
    else:
        abort(400)


# Add transaction
@app.route('/api/v1/addTransaction', methods=['POST'])
def addTransaction():
    if not request.json or not 'iban' in request.json:
        abort(400)
    db = client['defiexchange']
    coll = db['transactions']
    coll.insert_one(request.json)
    return jsonify({'success': "true"}), 201

@app.route("/api/swagger.json")
def create_swagger_spec():
    return jsonify(spec.to_dict())

app.run()