# src/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "defiexchange",
    "version": "1.0",
    "description": "API from DeFi Exchange to get user information and account history",
    "summary": "DeFi Exchange API",
    "contact": {
      "name": "defiexchange-support",
      "email": "suppert@defiexchange.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
  "paths": {
    "/accountHistory": {
      "parameters": [],
      "get": {
        "summary": "Get User Info by User ID",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "get-accountHistory",
        "description": "Get account history",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "legacyAddress",
            "description": "Legacy address of wallet",
            "required": true
          }
        ]
      }
    },
    "/userInformation": {
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "get-userInformation",
        "description": "Get user information",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "legacyAddress",
            "description": "Legacy address of Wallet",
            "required": true
          }
        ]
      }
    },
    "/addTransaction": {
      "post": {
        "summary": "",
        "tags": [],
        "operationId": "post-addTransaction",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Add new transaction\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "description": "",
                "type": "object",
                "x-examples": {
                  "example-1": {
                    "address": "81234",
                    "iban": "DE1234",
                    "signature": "sign",
                    "asset": "DFI",
                    "amount": "1.0234",
                    "timestamp": "2021-05-01 18:00"
                  }
                },
                "properties": {
                  "address": {
                    "type": "string",
                    "minLength": 1
                  },
                  "iban": {
                    "type": "string",
                    "minLength": 1
                  },
                  "signature": {
                    "type": "string",
                    "minLength": 1
                  },
                  "asset": {
                    "type": "string",
                    "minLength": 1
                  },
                  "amount": {
                    "type": "string",
                    "minLength": 1
                  },
                  "timestamp": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "address",
                  "iban",
                  "signature",
                  "asset",
                  "amount",
                  "timestamp"
                ]
              }
            }
          }
        }
      }
    }
  }
}